refact(na): move findPort and tryParseInt from utils file to separate utils files.
